+++ pkgs/checks/lib/src/extensions/math.dart
@@ -6,83 +6,101 @@ import '../../context.dart';
 
 extension NumChecks on Subject<num> {
   /// Expects that [num.isNaN] is true.
-  void isNaN() {
+  void get isNaN {
     context.expect(
       () => ['is not a number (NaN)'],
+      predicateNoun: () => 'NaN',
       (actual) => actual.isNaN ? null : Rejection(),
     );
   }
 
   /// Expects that [num.isNaN] is false.
-  void isNotNaN() {
+  void get isNotNaN {
     context.expect(
       () => ['is a number (not NaN)'],
+      predicateNoun: () => 'a number (not NaN)',
       (actual) => actual.isNaN ? Rejection() : null,
     );
   }
 
   /// Expects that [num.isNegative] is true.
-  void isNegative() {
-    context.expect(() => ['is negative'], (actual) {
-      if (actual.isNegative) return null;
-      return Rejection(which: ['is not negative']);
-    });
+  void get isNegative {
+    context.expect(
+      () => ['is negative'],
+      predicateNoun: () => 'a negative number',
+      (actual) =>
+          actual.isNegative ? null : Rejection(which: ['is not negative']),
+    );
   }
 
   /// Expects that [num.isNegative] is false.
-  void isNotNegative() {
-    context.expect(() => ['is not negative'], (actual) {
-      if (!actual.isNegative) return null;
-      return Rejection(which: ['is negative']);
-    });
+  void get isNotNegative {
+    context.expect(
+      () => ['is not negative'],
+      predicateNoun: () => 'a non-negative number',
+      (actual) => actual.isNegative ? Rejection(which: ['is negative']) : null,
+    );
   }
 
   /// Expects that [num.isFinite] is true.
-  void isFinite() {
-    context.expect(() => ['is finite'], (actual) {
-      if (actual.isFinite) return null;
-      return Rejection(which: ['is not finite']);
-    });
+  void get isFinite {
+    context.expect(
+      () => ['is finite'],
+      predicateNoun: () => 'a finite number',
+      (actual) => actual.isFinite ? null : Rejection(which: ['is not finite']),
+    );
   }
 
   /// Expects that [num.isFinite] is false.
   ///
   /// Satisfied by [double.nan], [double.infinity] and
   /// [double.negativeInfinity].
-  void isNotFinite() {
-    context.expect(() => ['is not finite'], (actual) {
-      if (!actual.isFinite) return null;
-      return Rejection(which: ['is finite']);
-    });
+  void get isNotFinite {
+    context.expect(
+      () => ['is not finite'],
+      predicateNoun: () => 'a non-finite number',
+      (actual) => actual.isFinite ? Rejection(which: ['is finite']) : null,
+    );
   }
 
   /// Expects that [num.isInfinite] is true.
   ///
   /// Satisfied by [double.infinity] and [double.negativeInfinity].
-  void isInfinite() {
-    context.expect(() => ['is infinite'], (actual) {
-      if (actual.isInfinite) return null;
-      return Rejection(which: ['is not infinite']);
-    });
+  void get isInfinite {
+    context.expect(
+      () => ['is infinite'],
+      predicateNoun: () => 'an infinite number',
+      (actual) =>
+          actual.isInfinite ? null : Rejection(which: ['is not infinite']),
+    );
   }
 
   /// Expects that [num.isInfinite] is false.
   ///
   /// Satisfied by [double.nan] and finite numbers.
-  void isNotInfinite() {
-    context.expect(() => ['is not infinite'], (actual) {
-      if (!actual.isInfinite) return null;
-      return Rejection(which: ['is infinite']);
-    });
+  void get isNotInfinite {
+    context.expect(
+      () => ['is not infinite'],
+      predicateNoun: () => 'a non-infinite number',
+      (actual) => actual.isInfinite ? Rejection(which: ['is infinite']) : null,
+    );
   }
 
   /// Expects that the difference between this number and [other] is less than
   /// or equal to [delta].
   void isCloseTo(num other, num delta) {
-    context.expect(() => ['is within <$delta> of <$other>'], (actual) {
-      final difference = (other - actual).abs();
-      if (difference <= delta) return null;
-      return Rejection(which: ['differs by <$difference>']);
-    });
+    context.expect(
+      () => ['is within <$delta> of <$other>'],
+      predicateNoun: () =>
+          switch ((literal(other).singleOrNull, literal(delta).singleOrNull)) {
+            (final other?, final delta?) => 'a value within $delta of $other',
+            _ => null,
+          },
+      (actual) {
+        final difference = (other - actual).abs();
+        if (difference <= delta) return null;
+        return Rejection(which: ['differs by <$difference>']);
+      },
+    );
   }
 }