+++ pkgs/checks/test/extensions/function_test.dart
@@ -15,7 +15,7 @@ void main() {
       });
       test('fails for functions that return normally', () {
         check(() {}).isRejectedBy(
-          (it) => it.throws<StateError>(),
+          .it()..throws<StateError>(),
           actual: ['a function that returned <null>'],
           which: ['did not throw'],
         );
@@ -27,13 +27,35 @@ void main() {
             StackTrace.fromString('fake trace'),
           );
         }).isRejectedBy(
-          (it) => it.throws<ArgumentError>(),
+          .it()..throws<ArgumentError>(),
           actual: ['a function that threw error <Bad state: oops!>'],
           which: [
             'threw an exception that is not a ArgumentError at:',
             '  fake trace',
           ],
         );
+      });
+      test('can be described', () {
+        check(
+          Condition.it<void Function()>()..throws<Object>(),
+        ).hasSyncDescription().deepEquals(['  throws an error']);
+        check(Condition.it<void Function()>()..throws<StateError>())
+            .hasSyncDescription()
+            .deepEquals(['  throws an error of type StateError']);
+      });
+      test('evaluates condition', () {
+        check(() => throw StateError('oops!')).isRejectedBy(
+          .it()..throws<StateError>(
+            .it()..has((e) => e.message, 'message').equals('wrong'),
+          ),
+          actual: ["'oops!'"],
+          which: ['differs at offset 0:', '  wrong', '  oops!', '  ^'],
+        );
+      });
+      test('returns valid subject', () {
+        check(
+          () => throw StateError('oops!'),
+        ).throws<StateError>().has((e) => e.message, 'message').equals('oops!');
       });
     });
 
@@ -48,10 +70,20 @@ void main() {
             StackTrace.fromString('fake trace'),
           );
         }).isRejectedBy(
-          (it) => it.returnsNormally(),
+          .it()..returnsNormally(),
           actual: ['a function that throws'],
           which: ['threw <Bad state: oops!> at:', '  fake trace'],
         );
+      });
+      test('evaluates condition', () {
+        check(() => 1).isRejectedBy(
+          .it()..returnsNormally(.it()..equals(2)),
+          actual: ['<1>'],
+          which: ['is not equal'],
+        );
+      });
+      test('returns valid subject', () {
+        check(() => 1).returnsNormally().equals(1);
       });
     });
   });