+++ pkgs/test_core/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.6.20-wip
+## 0.6.20
 
 * Support `globalSetup` and `addGlobalTearDown` host runner execution and
   caching across test suites.
@@ -13,17 +13,20 @@
 * Migrate out of process VM tests communication channel from TCP sockets to Unix
   domain sockets inside a restricted temporary directory.
 * Enable asserts in the `exe` compiler.
+* Fix deadlock that may occur running tests with the `exe` compiler.
 
 ## 0.6.19
 
 * Add support for `-c cli` (the native CLI compiler) to the vm platform.
 * Support using the OS platform selector to configure browser tests.
 * Use a DevTools URL instead of a defunct observatory URL.
+* Add flag `--shard-by-suite` to control sharding strategy.
 * Allow package_config `3.x.x`.
 * Require `analyzer: '>=13.0.0 <15.0.0'`
 * Update `parse_metadata.dart` to be compatible with `analyzer >=13.0.0 <15.0.0`.
 * Use the compact or failures-only reporters by default for tests run directly
   instead of through the test runner.
+
 
 ## 0.6.18